# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Heymaty - OpenAPI 3.0",
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://heymaty.com/"
    }
  ],
...
  "paths": {
...
    "/language/": {
      "get": {
        "tags": [
          "language"
        ],
        "summary": "Get Languages route",
        "description": "Get Languages route",
        "operationId": "getLanguages",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "languages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "64ff04fdbba02bede6a5fc24"
                          },
                          "name": {
                            "type": "string",
                            "example": "Ruzke"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Все языки"
                    },
                    "length": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl",
          "user"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServerErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "msg": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "errors"
        ]
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}